+++ tests/conftest.py
@@ -1,16 +1,13 @@
 import logging
 import os
-import sys
-from unittest.mock import patch
 
 import pandas as pd
-import pyarrow
 import pytest
 from _pytest.logging import caplog as _caplog  # noqa F401
 from databricks import koalas as ks
 from loguru import logger
 
-from eds_scikit.io.improve_performance import load_koalas, improve_performances, set_env_variables
+from eds_scikit.io.improve_performance import improve_performances
 
 from . import test_registry  # noqa: F401 --> To register functions
 
@@ -128,11 +125,3 @@ def example_objects():
             ks.Series([10, 11, 12]),
         ],
     )
-
-
-def test_improve_performances():
-    del sys.modules["databricks.koalas"]
-    load_koalas()
-
-    with patch.object(pyarrow, "__version__", "2.1.0"):
-        set_env_variables()
